Split fleet.py monolith into fleet/ package for SRP
460-line God Object decomposed into 7 focused modules:
- models.py: dataclasses
- loader.py: YAML parsing & validation
- assignment.py: host assignment & stable seeding
- orchestrator.py: archive generation
- manifest.py: fleet.manifest writer
- warnings.py: override conflict detection
- display.py: dry-run output

__init__.py re-exports all public symbols for backwards compatibility.
All 510 tests pass, no consumer changes needed.
